component version of dashboard
