+++ frontend/src/lib/types.ts
@@ -1,0 +1,9 @@
+export interface Recipe {
+  id: number;
+  title: string;
+  description: string;
+  ingredients: string[];
+  steps: string[];
+  tags: string[];
+  votes: number;
+}
